requirements.txt
streamlit
pandas
numpy
plotly
folium
streamlit-folium
fpdf2
kaleido
geopy
unidecode